# lib/api-zod/src/generated/types/createAnalysisRecordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Design Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAnalysisRecordBody {
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  fileId?: number | null;
  title: string;
  variationType: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  parameters?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  referenceFiles?: string | null;
  isFavorite: boolean;
}