# lib/api-zod/src/generated/types/fileRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Design Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface FileRecord {
  id: number;
  /** @nullable */
  projectId?: number | null;
  fileName: string;
  fileType: string;
  /** @nullable */
  fileSizeKb?: number | null;
  /** @nullable */
  bridgeType?: string | null;
  /** @nullable */
  spanLength?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  loadCapacity?: number | null;
  /** @nullable */
  designCode?: string | null;
  /** @nullable */
  layers?: string | null;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  extractedData?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  notes?: string | null;
  analysisStatus: string;
  createdAt: Date;
  updatedAt: Date;
}