# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Design Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { CountByLabel } from "./countByLabel";
import type { FileRecord } from "./fileRecord";

export interface StatsSummary {
  totalProjects: number;
  totalFiles: number;
  totalRecords: number;
  totalComparisons: number;
  filesByType: CountByLabel[];
  recordsByVariationType: CountByLabel[];
  recentFiles: FileRecord[];
}