# lib/api-zod/src/generated/types/variationPattern.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bridge Design Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisRecord } from "./analysisRecord";

export interface VariationPattern {
  variationType: string;
  count: number;
  examples: AnalysisRecord[];
}